+++ pypeit/images/combineimage.py
@@ -174,7 +174,6 @@ class CombineImage(object):
             for ff, file in enumerate(self.files):
                 print(msgs.indent() + strout.format(os.path.split(file)[1], " ".join(lampstat[ff].split("_"))))
             print(msgs.indent() + '-'*maxlen + "  " + '-'*maxlmp)
-            embed(header='')
 
         # Coadd them
         weights = np.ones(nimages)/float(nimages)