+++ pypeit/spectrographs/__init__.py
@@ -18,5 +18,6 @@ from pypeit.spectrographs import vlt_xshooter
 from pypeit.spectrographs import wht_isis
 from pypeit.spectrographs import lbt_mods
 from pypeit.spectrographs import vlt_fors
+from pypeit.spectrographs import mdm_osmos
 from pypeit.spectrographs import lbt_luci
 from pypeit.spectrographs import mmt_binospec